users can now publish and unpublish a did to the api
