+++ stores/settings.js
@@ -1,5 +1,4 @@
 import { atomWithStorage } from "jotai/utils";
 import { getRandomAvatar } from "../utils/contacts";
-import localforage from "localforage";
 
 export const myAvatarAtom = atomWithStorage("myAvatar", getRandomAvatar());
